labs/lab-21-python-confluence-publisher/requirements.txt
atlassian-python-api==3.41.14
python-dotenv==1.0.1
markdown==3.5.2
